Refactor: Extract sync dialog components and add multi-platform design
- Remove unused isOpen from SyncState interface
- Extract shared types to sync/types.ts (DraftSummary, MergeConfig, etc.)
- Extract NewDesignPreviewDialog to separate file
- Extract MergePreviewDialog to separate file
- Add PLATFORM_SYNC_DESIGN.md documenting abstraction plan for
  Printables and Thingiverse support
- Reduce makerworld-sync.tsx from 2142 to 1736 lines
