+++ src/app/components/dashboard/sync/index.ts
@@ -1,0 +1,3 @@
+export * from "./types";
+export { NewDesignPreviewDialog } from "./new-design-preview-dialog";
+export { MergePreviewDialog } from "./merge-preview-dialog";
